bigint

+++ db/migrate/20240115183557_create_notifications.rb
@@ -1,11 +1,11 @@
 class CreateNotifications < ActiveRecord::Migration[6.1]
   def change
     create_table :notifications do |t|
-      t.integer :visitor_id, null: false
-      t.integer :visited_id, null: false
-      t.integer :comment_id, null: false
-      t.integer :like_id, null: false
-      t.integer :follower_id, null: false
+      t.bigint :visitor_id, null: false
+      t.bigint :visited_id, null: false
+      t.bigint :comment_id, null: false
+      t.bigint :like_id, null: false
+      t.bigint :follower_id, null: false
       t.string :action, null: false
       t.boolean :is_checked, null: false, default: false
