+++ db/migrate/20240115183705_create_follows.rb
@@ -1,8 +1,8 @@
 class CreateFollows < ActiveRecord::Migration[6.1]
   def change
     create_table :follows do |t|
-      t.integer :follower_id, null: false
-      t.integer :followee_id, null: false
+      t.bigint :follower_id, null: false
+      t.bigint :followee_id, null: false
 
       t.timestamps
     end
