+++ db/migrate/20240212084449_create_relationships.rb
@@ -1,8 +1,8 @@
 class CreateRelationships < ActiveRecord::Migration[6.1]
   def change
     create_table :relationships do |t|
-      t.integer :follower_id
-      t.integer :followed_id
+      t.bigint :follower_id
+      t.bigint :followed_id
 
       t.timestamps
     end
