+++ db/migrate/20240223171419_add_post_id_to_notifications.rb
@@ -1,5 +1,5 @@
 class AddPostIdToNotifications < ActiveRecord::Migration[6.1]
   def change
-    add_column :notifications, :post_id, :integer
+    add_column :notifications, :post_id, :bigint
   end
 end
